# lib/api-zod/src/generated/types/extractedIntent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Canonical project intent. Select fields (organizationType, tone, callToAction, technologyStack) are constrained to the form's allowed values or an empty string when unknown.

 */
export interface ExtractedIntent {
  projectName: string;
  businessName: string;
  founderName: string;
  organizationType: string;
  primaryGoal: string;
  audience: string;
  services: string[];
  pages: string[];
  tone: string;
  callToAction: string;
  technologyStack: string;
  contactEmail: string;
  contactPhone: string;
  notes: string;
}