# lib/api-zod/src/generated/types/mineIntentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MineIntentRequest {
  /**
   * The raw brain-dump text to mine.
   * @minLength 1
   * @maxLength 8000
   */
  text: string;
}