# lib/api-zod/src/generated/types/minedIntentResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedIntent } from "./extractedIntent";
import type { MinedIntentResultProjectKind } from "./minedIntentResultProjectKind";
import type { MinedIntentResultSource } from "./minedIntentResultSource";

export interface MinedIntentResult {
  intent: ExtractedIntent;
  projectKind: MinedIntentResultProjectKind;
  filledFields: string[];
  missingFields: string[];
  suggestions: string[];
  source: MinedIntentResultSource;
}